requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
requests==2.28.2
Pillow==9.5.0
tenacity==8.2.2
PyYAML==6.0
